requirements.txt
environs==9.5.0
requests==2.28.2
redis==4.5.1
python-telegram-bot==13.*
